Refactored Dialog box from Purchasing view so address and phone input components may be reused
