+++ src/main/app/src/components/Purchasing.js
@@ -6,14 +6,9 @@ import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import TextField from '@material-ui/core/TextField';
-import MaskedInput from 'react-text-mask';
-import Input from '@material-ui/core/Input';
 import InputLabel from '@material-ui/core/InputLabel';
-import FormControl from '@material-ui/core/FormControl';
-import PropTypes from 'prop-types';
-import MenuItem from '@material-ui/core/MenuItem';
-import Select from '@material-ui/core/Select';
-import Grid from '@material-ui/core/Grid';
+import AddressInput from '../components/forms/AddressInput';
+import PhoneInput from '../components/forms/PhoneInput';
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -53,82 +48,6 @@ const orderRows = [
   { id: 4, vendorName: 'Vendor1', items: '????', timestamp: "01/31/2021", cost: "$100" }
 ];
 
-function TextMaskCustom(props) {
-  const { inputRef, ...other } = props;
-
-  return (
-    <MaskedInput
-      {...other}
-      ref={(ref) => {
-        inputRef(ref ? ref.inputElement : null);
-      }}
-      mask={['(', /[1-9]/, /\d/, /\d/, ')', ' ', /\d/, /\d/, /\d/, '-', /\d/, /\d/, /\d/, /\d/]}
-      placeholderChar={'\u2000'}
-      showMask
-    />
-  );
-}
-
-TextMaskCustom.propTypes = {
-  inputRef: PropTypes.func.isRequired,
-};
-
-const PhoneNumberField = () => {
-  const [values, setValues] = React.useState({
-    textmask: '(  )    -    ',
-    numberformat: '1320',
-  });
-
-  const handleChange = (event) => {
-    setValues({
-      ...values,
-      [event.target.name]: event.target.value,
-    });
-  };
-  return (
-    <FormControl>
-      <InputLabel>Phone Number</InputLabel>
-      <Input
-        value={values.textmask}
-        onChange={handleChange}
-        name="textmask"
-        inputComponent={TextMaskCustom}
-      />
-    </FormControl>
-  );
-};
-
-const ProvinceDrawer = () => {
-  const [province, setProvince] = React.useState('');
-
-  const handleChange = (event) => {
-    setProvince(event.target.value);
-  };
-  return (
-    <div >
-      <FormControl>
-        <Select
-          value={province}
-          onChange={handleChange}
-          margin="normal"
-        >
-          {
-            ['AB', 'BC', 'MB',
-              'NB', 'NL', 'NT',
-              'NS', 'NU', 'ON',
-              'PE', 'QC', 'SK', 'YT'
-            ].map(province =>
-              <MenuItem value={province}>
-                {province}
-              </MenuItem>
-            )
-          }
-        </Select>
-      </FormControl>
-    </div>
-  );
-};
-
 function LoadedView() {
   const [open, setOpen] = React.useState(false);
 
@@ -154,34 +73,8 @@ function LoadedView() {
             label="Required"
             type="text"
           />
-          <InputLabel>Address</InputLabel>
-          <TextField
-            autoFocus
-            margin="normal"
-            type="text"
-            fullWidth />
-          <Grid container spacing={2}>
-            <Grid item md={3}>
-              <InputLabel>City</InputLabel>
-              <TextField
-                autoFocus
-                margin="normal"
-                type="text" />
-            </Grid>
-            <Grid item md={3}>
-              <InputLabel>Postal Code</InputLabel>
-              <TextField
-                autoFocus
-                margin="normal"
-                inputProps={{ maxLength: 6 }}
-                type="text" />
-            </Grid>
-            <Grid item md={3}>
-              <InputLabel>Province</InputLabel>
-              <ProvinceDrawer />
-            </Grid>
-          </Grid>
-          <PhoneNumberField />
+          <AddressInput />
+          <PhoneInput />
         </DialogContent>
         <DialogActions>
           <Button onClick={handleClose} color="primary">
